backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
black==25.1.0
boto3==1.37.33
botocore==1.37.33
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.2
paramiko==3.5.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sshtunnel==0.4.0
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
